Moved located vector client class to cpp file
